refactor: extract useChatSearch + expand useMessageSelection with delete/hide actions
